Today's gospel speaks to (dec|)centralization
There's beautiful wisdom in today's episode of Love's life.

Us 13 only have a little fermented food and a couple fish.

Well, what could other such groups have in 5000, right?

So do we gather everyone under one hierarchy, expropriate
their bread and share equally among all? Nope.

Do we make it everyone for themself and let those without
starve? Not that either.

What type of language are we dealing with? Apes.

What can be counted on to be in their standard library?
Trust-based resource sharing.

What's the implementation's upper bound?
200 tops.

We're in a constrained situation, so make it fifty.

And there was abundance.

A big network of little trust hierarchies is manageable
and noncoercive.

Do we centralize the nets? Do we decentralize the nets?
Wrong language.

Let us be the reforestation who we want to see in the nets.

Amen.